Introduce OAuth reducer for the OAuth related actions
